Task/requirements.txt
mariadb
python-dotenv
requests
python-dateutil